Frontporting bug fixes to 3.2 dev (#1765)
Fixes for #1770 #1760 #1356 on 3.2
* [admin] proting fixes from master
* [locale] Updating locale changes
* [doc] updating release notes with fixed issues
* [tsp][C++] fixing for 20.04
* [tsp][pgtap] testing issue
* [doc] updating NEWS and full version results
* Porting Fixes for Issue 1770 on 3.2
* [CI] adding action for clang compiler
* [C] Ingnoring postgres warnings
* [dijkstra] compiling with clang
* [common] compiling with clang
* [astar] compiles with clang
* [ksp] compiles with clang
* [tsp] compiles with clang
* [alpha_shape] compiles with clang
* [tspEuclidean] compiles with clang
* [witPoints] compiles with clang
* [bellman_ford] compiles with clang
* [chinese] compiles with clang
* [transitiveClosure] compiles with clang
* [breadthFirstSearch] compiles with clang
* [components] compiles with clang
* [maxflow] pgr_maxflow.hpp compiles with clang
* [maxflow] compiles with clang
* [lineGraph] compiles with clang
* [bdAstar] compiles with clang
* [bdDijkstra] compiles with clang
* [trsp] compiles with clang
* [pickDeliver] compiles with clang
* [dijkstra][docqueries] giving a new correct alternate result
* Removing a test of a function that does not exist on pgRouting
* removing internal query tests that were created when developing contraction and are not used
* [tsp][docqueries] Modification of rand() gives new results
* roll back changes that affect ksp
* [C++] moving pgr_messages.cpp to cpp_common
* roll back changes that affect trsp
* src/common/basePath_SSEC.cpp compiles with clang
* src/trsp/pgr_trspHandler.cpp compiles with clang
* [ci][pgtap] adding testing after clang compilation
* [doc] adding NEWS & Release notes
* Also updating compilation_time and compiler & hash on pgr_full_version
* [build] rolling back build-extension-update-files1.pl
* [C++] adding some minimal changes from the comments of the PR
* Fixing merge conflicts
* [clang] Some additional fixes for clang compilation
* [doc] updating full version execution for this PR
* [locale] updating latest documentation for translations

+++ include/chinese/pgr_chinesePostman.hpp
@@ -90,7 +90,7 @@ class PgrDirectedChPPGraph {
      /** vector of vertex -> vector of edges **/
      std::vector<std::pair<int64_t, std::vector<size_t>>> resultGraph;
      std::map<int64_t, size_t> VToVecid;
-     Identifiers<int64_t> edgeVisited;
+     Identifiers<size_t> edgeVisited;
      Identifiers<int64_t> vertexVisited;
 
      std::stack<int64_t> pathStack;  // node stack
